Added explicit cast to int to prevent sign extension of ioctl request identifiers (apparently sizeof() is 8-byte long on x86-64).
git-svn-id: https://svn.aros.org/svn/aros/trunk@28691 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/include/sys/ioctl.h
@@ -66,7 +66,7 @@ __END_DECLS
 #define	IOC_DIRMASK	0xe0000000	             /* mask for IN/OUT/VOID */
 
 #define _IOC(inout,group,num,len) \
-	(inout | ((len & IOCPARM_MASK) << 16) | ((group) << 8) | (num))
+	(inout | (((int)(len) & IOCPARM_MASK) << 16) | ((group) << 8) | (num))
 #define	_IO(g,n)	_IOC(IOC_VOID,	(g), (n), 0)
 #define	_IOR(g,n,t)	_IOC(IOC_OUT,	(g), (n), sizeof(t))
 #define	_IOW(g,n,t)	_IOC(IOC_IN,	(g), (n), sizeof(t))
